+++ apps/bare-expo/ios/BareExpo/AppDelegate.swift
@@ -18,7 +18,7 @@ import FlipperKit
 #endif
 
 @UIApplicationMain
-class AppDelegate: ExpoAppDelegate {
+class AppDelegate: ExpoAppDelegate, RCTBridgeDelegate, EXDevLauncherControllerDelegate {
   var bridge: RCTBridge?
   var launchOptions: [UIApplication.LaunchOptionsKey: Any]?
 
@@ -42,19 +42,17 @@ class AppDelegate: ExpoAppDelegate {
   }
 
   @discardableResult
-  func initializeReactNativeBridge(_ launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> RCTBridge? {
-    if let bridge = RCTBridge(delegate: self, launchOptions: launchOptions) {
-      let rootView = RCTRootView(bridge: bridge, moduleName: "main", initialProperties: nil)
-      let rootViewController = EXScreenOrientationViewController()!
-      rootView.backgroundColor = UIColor.white
-      rootViewController.view = rootView
+  func initializeReactNativeBridge(_ launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> RCTBridge {
+    let bridge = reactDelegate.createBridge(delegate: self, launchOptions: launchOptions)
+    let rootView = reactDelegate.createRootView(bridge: bridge, moduleName: "main", initialProperties: nil)
+    let rootViewController = reactDelegate.createRootViewController()
+    rootView.backgroundColor = UIColor.white
+    rootViewController.view = rootView
 
-      window?.rootViewController = rootViewController
-      window?.makeKeyAndVisible()
-      self.bridge = bridge
-      return bridge;
-    }
-    return nil;
+    window?.rootViewController = rootViewController
+    window?.makeKeyAndVisible()
+    self.bridge = bridge
+    return bridge
   }
   
   #if RCT_DEV
@@ -82,11 +80,9 @@ class AppDelegate: ExpoAppDelegate {
     client?.start()
   #endif
   }
-}
 
-// MARK: - RCTBridgeDelegate
+  // MARK: - RCTBridgeDelegate
 
-extension AppDelegate: RCTBridgeDelegate {
   func sourceURL(for bridge: RCTBridge!) -> URL! {
     // DEBUG must be setup in Swift projects: https://stackoverflow.com/a/24112024/4047926
     #if DEBUG
@@ -99,7 +95,7 @@ extension AppDelegate: RCTBridgeDelegate {
     return Bundle.main.url(forResource: "main", withExtension: "jsbundle")
     #endif
   }
-  
+
   func extraModules(for bridge: RCTBridge!) -> [RCTBridgeModule] {
     var extraModules = [RCTBridgeModule]()
     // You can inject any extra modules that you would like here, more information at:
@@ -109,7 +105,7 @@ extension AppDelegate: RCTBridgeDelegate {
     // https://github.com/expo/react-native/commit/7f2912e8005ea6e81c45935241081153b822b988
     // Let's bring it back in Bare Expo.
     extraModules.append(RCTDevMenu() as! RCTBridgeModule)
-    
+
     // Add AsyncStorage back to the project
     // https://github.com/expo/react-native/commit/bd1396034319e6e59f960fac7aeca1f483c2052d
     let documentDirectory = NSSearchPathForDirectoriesInDomains(.documentDirectory, .userDomainMask, true).first! as NSString
@@ -117,11 +113,9 @@ extension AppDelegate: RCTBridgeDelegate {
     extraModules.append(RCTAsyncLocalStorage(storageDirectory: storageDirectory))
     return extraModules
   }
-}
 
-// MARK: - EXDevelopmentClientControllerDelegate
+  // MARK: - EXDevelopmentClientControllerDelegate
 
-extension AppDelegate:  EXDevLauncherControllerDelegate {
   func devLauncherController(_ developmentClientController: EXDevLauncherController, didStartWithSuccess success: Bool) {
     developmentClientController.appBridge = initializeReactNativeBridge(developmentClientController.getLaunchOptions())
   }