+++ packages/expo-ads-admob/src/PublisherBanner.web.tsx
@@ -1,12 +1,10 @@
 import * as React from 'react';
 import { Text, View } from 'react-native';
 
-export default class PublisherBanner extends React.Component<object> {
-  render() {
-    return (
-      <View>
-        <Text>PublisherBanner component not supported on the web</Text>
-      </View>
-    );
-  }
-}
+const PublisherBanner = () => (
+  <View>
+    <Text>PublisherBanner component not supported on the web</Text>
+  </View>
+);
+
+export default PublisherBanner;