+++ packages/expo-module-template/src/ExpoModuleTemplateNativeView.ts
@@ -1,4 +1,5 @@
 import { requireNativeViewManager } from 'expo-modules-core';
+import * as React from 'react';
 
 type NativeViewProps = {
   someGreatProp: number;