+++ packages/expo-modules-core/ios/Swift/ModulesProvider.swift
@@ -18,6 +18,12 @@ public protocol ModulesProviderProtocol: ModulesProviderObjCProtocol {
    Returns an array of classes that hooks into `ExpoAppDelegate` to receive app delegate events.
    */
   func getAppDelegateSubscribers() -> [ExpoAppDelegateSubscriber.Type]
+
+  typealias ExpoReactDelegateHandlerTupleType = (packageName: String, handler: ExpoReactDelegateHandler.Type)
+  /**
+   Returns an array of `ExpoReactDelegateHandlerTupleType` for `ReactDelegate` to hook React instance creation.
+   */
+  func getReactDelegateHandlers() -> [ExpoReactDelegateHandlerTupleType]
 }
 
 /**
@@ -31,6 +37,10 @@ open class ModulesProvider: NSObject, ModulesProviderProtocol, ModulesProviderOb
   }
 
   open func getAppDelegateSubscribers() -> [ExpoAppDelegateSubscriber.Type] {
+    return []
+  }
+
+  open func getReactDelegateHandlers() -> [ExpoReactDelegateHandlerTupleType] {
     return []
   }
 }